fixed pictures DDL and data. changed photos.aspx to generate dynamically. created struct for photos data. renamed loginclass.cs to kaminariD
Added (currently filler) alt text to the pictures table and column in
DDL.
renamed loginclass to kaminariDB to be a more accurate name for the
class and it's functions.
created PhotoStruct.cs for holding photo name, description and alt text.
updated login.aspx to use new class name.
modified photos.aspx to generate dynamically based on photos in
database.